+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 ## Improvements
 - Improved the iCloud Sync logic to only copy the data from the cloud, rather than using it as a source directly. This translates in better performances and less possibility to launch a game that is not actually downloaded (if you had no data for example)
 
+## Features
+- Added a share screenshot button in the overlay menu
+- Added a save screenshot button in the overlay menu
+
 # 1.0
 
 Initial release